Integrate VerifySignature into one at server/gateways/ethereum.
