+++ infrastructure/function_scripts/cmd/verify_auth_challenge_response/main.go
@@ -2,15 +2,12 @@ package main
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
-	"github.com/ethereum/go-ethereum/accounts"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/team-azb/knowtfolio/infrastructure/function_scripts/pkg/aws_utils"
+	"github.com/team-azb/knowtfolio/server/gateways/ethereum"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
@@ -32,38 +29,11 @@ func handler(ctx context.Context, event *events.CognitoEventUserPoolsVerifyAuthC
 	sign := event.Request.ChallengeAnswer.(string)
 	signedData := event.Request.PrivateChallengeParameters["sign_message"]
 
-	err = verifySignature(address, sign, signedData)
+	err = ethereum.VerifySignature(address, sign, signedData)
 	if err == nil {
 		event.Response = events.CognitoEventUserPoolsVerifyAuthChallengeResponse{AnswerCorrect: true}
 	}
 	return event, nil
-}
-
-// verifySignature checks if `sign` is a signature that is signed by `addr` using `signedData`.
-func verifySignature(addr string, sign string, signedData string) error {
-	decodedSign, err := hexutil.Decode(sign)
-	if err != nil {
-		return err
-	}
-
-	// NOTE: Some client generates 27/28 instead of 0/1 due to Legacy Ethereum
-	// ref) https://github.com/ethereum/go-ethereum/issues/19751#issuecomment-504900739
-	if decodedSign[crypto.RecoveryIDOffset] >= 27 {
-		decodedSign[crypto.RecoveryIDOffset] -= 27
-	}
-
-	hash := accounts.TextHash([]byte(signedData))
-	pubKey, err := crypto.SigToPub(hash, decodedSign)
-	if err != nil {
-		return err
-	}
-
-	signedAddr := crypto.PubkeyToAddress(*pubKey)
-	if addr != signedAddr.String() {
-		return errors.New("`address` didn't match the signer")
-	}
-
-	return nil
 }
 
 func main() {
