+++ server/gateways/ethereum/verify_signature.go
@@ -1,0 +1,35 @@
+package ethereum
+
+import (
+	"errors"
+	"github.com/ethereum/go-ethereum/accounts"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
+// VerifySignature checks if `sign` is a signature that is signed by `addr` using `signedData`.
+func VerifySignature(addr string, sign string, signedData string) error {
+	decodedSign, err := hexutil.Decode(sign)
+	if err != nil {
+		return err
+	}
+
+	// NOTE: Some client generates 27/28 instead of 0/1 due to Legacy Ethereum
+	// ref) https://github.com/ethereum/go-ethereum/issues/19751#issuecomment-504900739
+	if decodedSign[crypto.RecoveryIDOffset] >= 27 {
+		decodedSign[crypto.RecoveryIDOffset] -= 27
+	}
+
+	hash := accounts.TextHash([]byte(signedData))
+	pubKey, err := crypto.SigToPub(hash, decodedSign)
+	if err != nil {
+		return err
+	}
+
+	signedAddr := crypto.PubkeyToAddress(*pubKey)
+	if addr != signedAddr.String() {
+		return errors.New("`address` didn't match the signer")
+	}
+
+	return nil
+}
